feat(gates): let one gate have two provers without either faking a proof
`snyk` and `npm_security_scan` both prove `dependency-vulnerability`, and only
one of them read the declaration. Declaring the gate `off` silenced the audit
and left the supply-chain scan running — a declaration satisfied in one job and
ignored in another.

The hazard here is a PASS, which is what makes it different from the precedent.
`gates-and-policy.md` works through one prover and two gates, where a FAILURE
risks being attributed to a property nobody measured. This is the transpose,
and its failure mode is quieter: `🔒 Security Scan` is a required context here
and can be green while the second scanner examined nothing, because it declines
silently whenever SNYK_TOKEN is absent.

Three decisions, taken and encoded rather than left to a comment.

WHICH JOB CARRIES THE CONTEXT. `npm_security_scan`, which already posts the
gate's label and is the required check. `snyk` is now a declared SECOND prover:
`QUALITY_JOB_GATES` gains a `snyk` row for the same gate — the table is keyed by
job precisely so that is expressible — while the LABEL stays singular. The
façade fixture models this as `secondaryProver`, and the suite asserts both
halves: a secondary must NOT be named the label, and every gate has exactly one
job that is. That last assertion is new and is the one that bites hardest: two
jobs agreeing on the label is the failure, and agreeing is what makes it
invisible.

WHAT A DECLARATION MEANS. Both provers, coherently. `snyk` resolves the gate
before anything else, so `off` stops both. On the configured path it does not
re-run the project's task — that runs once, in the job that carries the label —
it records the handover and stops. Proving one property twice under two names
is the thing a second prover must not do.

DECLINED IS NOT PROVED. The skip notice now names the property, says which half
went unproven and why, and says out loud that a green on the other job does not
mean this one ran. It is a ::warning an audit can key on rather than an echo
that reads as housekeeping.

Case 2 of the issue is obsolete and was verified so rather than assumed:
`test_e2e` was deleted in #2841, and nothing in this repository calls
playwright-e2e.yml, so there is no longer a second prover of `e2e-browser` to
disagree with the first.

`UNGATED_QUALITY_JOBS` drops to four. The façade suite split in two along the
seam it already had — identity in one file, mechanics in the other — because it
crossed the 300-line cap.

Work-Item: CodySwannGT/lisa#2915
