+++ tests/integration/quality-gate-facade-jobs.ts
@@ -60,6 +60,21 @@ export interface ConvertedJob {
   gateStep: string;
   fallbackSteps: string[];
   /**
+   * Whether this job is the SECOND prover of a gate another job carries.
+   *
+   * Absent means primary: the job's `name:` is the gate's `label`, and it is
+   * the job a ruleset matches. Present means the opposite is required — the
+   * job proves the same property at a different depth and must NOT be named
+   * the label, because two jobs posting one branch-protection context is a
+   * check nobody can reason about.
+   *
+   * There is exactly one of these today, and the suite asserts both halves:
+   * one primary per gate, and a secondary that does not wear the primary's
+   * name.
+   */
+  secondaryProver?: true;
+
+  /**
    * The workflow file that declares the job.
    *
    * Carried per entry rather than assumed, because the façade contract is no
@@ -362,6 +377,20 @@ const CONVERTED_JOBS: ConvertedJobSource[] = [
       "📊 SonarCloud Scan",
       "🔍 Validate SonarCloud results",
       "📊 SonarCloud Scan Skipped",
+    ],
+  },
+  {
+    job: "snyk",
+    jobName: "🛡️ Snyk Dependency Scan",
+    secondaryProver: true,
+    gateStep: "🛡️ dependency-vulnerability is proved by its own gate job",
+    // The whole vendor path, token probe included. A project that declared its
+    // own task is not using this scanner, so a step that announces a missing
+    // key for it would be reporting on tooling nobody asked for.
+    fallbackSteps: [
+      "🔍 Check for Snyk token",
+      "🛡️ Run Snyk to check for vulnerabilities",
+      "🛡️ Snyk Scan Skipped",
     ],
   },
   {
